Deprecation is now in the Concern module.
